requirements.txt
torch
torchvision
numpy
pandas
scikit-learn
matplotlib
wandb
tqdm
nibabel
scikit-image
Pillow
